change enumerators to a 'base' inheritance model like what we did for elements
