+++ lib/kitchen.rb
@@ -20,25 +20,6 @@ require "kitchen/utils"
 require "kitchen/errors"
 require "kitchen/ancestor"
 require "kitchen/search_history"
-require "kitchen/element_enumerator"
-require "kitchen/element_enumerator_factory"
-require "kitchen/page_element_enumerator"
-require "kitchen/term_element_enumerator"
-require "kitchen/chapter_element_enumerator"
-require "kitchen/book_element_enumerator"
-require "kitchen/figure_element_enumerator"
-require "kitchen/table_element_enumerator"
-require "kitchen/example_element_enumerator"
-require "kitchen/element_base"
-require "kitchen/element"
-require "kitchen/book_element"
-require "kitchen/chapter_element"
-require "kitchen/example_element"
-require "kitchen/page_element"
-require "kitchen/term_element"
-require "kitchen/figure_element"
-require "kitchen/table_element"
-require "kitchen/note_element"
 require "kitchen/document"
 require "kitchen/book_document"
 require "kitchen/debug/print_recipe_error"
@@ -48,8 +29,14 @@ require "kitchen/oven"
 require "kitchen/clipboard"
 require "kitchen/pantry"
 require "kitchen/counter"
-require "kitchen/note_element_enumerator"
 require "kitchen/patches"
+
+require "kitchen/element_enumerator_base"
+require "kitchen/element_enumerator_factory"
+require_all("kitchen", "*enumerator.rb")
+
+require "kitchen/element_base"
+require_all("kitchen", "*element.rb")
 
 require_all("kitchen/directions")
