+++ lib/kitchen/book_element_enumerator.rb
@@ -1,5 +1,5 @@
 module Kitchen
-  class BookElementEnumerator < ElementEnumerator
+  class BookElementEnumerator < ElementEnumeratorBase
     # Unsure if this class is needed; there is only one book element per document
   end
 end
