+++ lib/kitchen/example_element_enumerator.rb
@@ -1,5 +1,5 @@
 module Kitchen
-  class ExampleElementEnumerator < ElementEnumerator
+  class ExampleElementEnumerator < ElementEnumeratorBase
 
     def self.factory
       ElementEnumeratorFactory.new(
